SLING-10502 - fix package name typo (that no one caught so far ;-)
